expense-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose